create permission user

+++ controller/delete-todos.go
@@ -14,6 +14,19 @@ func DeleteTodos(e *echo.Echo, db *sql.DB) {
 		user := ctx.Get("USER").(models.AuthClaimJWT)
 		id := ctx.Param("id")
 
+		//set permission untuk user
+		permissionFound := false
+		for _, scope := range user.UserScopes {
+			if scope == "todos:delete" {
+				permissionFound = true
+				break
+			}
+		}
+
+		if !permissionFound {
+			return ctx.String(http.StatusForbidden, "Forbidden")
+		}
+
 		_, err := db.Exec(
 			"DELETE FROM todos WHERE id = ? AND user_id = ?",
 			id,
